==> notebooks/00_dataExploration.ipynb <==
import pandas as pd 
import numpy as np 

pd.options.display.max_rows=1000

DATA_PATH='../data/raw/survey_results_public.csv'

# Read data and print shape
raw_data=pd.read_csv(DATA_PATH)
raw_data.shape

# Display random answer
raw_data.sample(1).iloc[0]

# Observation: Multiple answers need to be splitted 
# Reference to the schema needed to understand

# Print general information of the data frame 
raw_data.info()

# Get stats for numerical columns
raw_data.describe() 

# Investigate the questionable objects columns
questionable_columns=['YearsCode','YearsCodePro']

for col in questionable_columns:
    print(col)
    print(raw_data[col].unique().tolist())
    print('_'*100)
    print()
