==> notebooks/01_Preprocessing.ipynb <==
# constants
DATA_PATH='../data/raw/survey_results_public.csv'
EXPORT_PATH = "../data/processed/1_preprocessed_data.pkl"

REPLACE_COLS=['YearsCode','YearsCodePro']
REPLACE_DICT={
    'Less than 1 year' : 0,
    'More than 50 years' : 51 
}

# packages
import pandas as pd 
import numpy as np
import logging
import pickle

# Functions

def Replace_nulls(pd_series,replacement=[]):
    '''
        Aim:
        Replace null values with a specific replacement
        
        Parmeters:
        * pd_series (pd.Series): String series with answers
        * replacement (list) : list to replace null value
        
        Return: (pd.Series): series after replacing 
    '''
    nulls=pd_series.isnull()
    pd_series.loc[nulls]=pd_series.loc[nulls].apply(lambda x: replacement)
    return pd_series

def split_multianswers(answers,delimit=';'):
    '''
        Aim:
        Split multiple answers in a single string 
        to a list of strings each represnting a single answer
        
        Parameters:
        * answers (pd.Series): String series with answers 
        * delimit (string): endpoint for each single answer Defaults to ";"
        
        Return: (pd.Series): If column contains multiple answers
    
    '''
     # Sub functions 
    def is_splittable(answer,delimit):
        """ Check if answer should be splitted - Returns boolean """    
        return answer.str.contains(delimit)
    
    def split_answer(answer,delimit):
        """Function to split single answer - Returns list of strings"""
        return answer.str.split(delimit)
    
    #__________________________________________________________#
    
    # Check if multiple answers exist - if none: return original 
    splittable_answers=is_splittable(answers,delimit)
    if not splittable_answers.any():
        return answers
    
    # Split each multiple answer to list of single answers
    splitted_answers=split_answer(answers,delimit)
    
    # Replace NAs with empty lists 
    modified_values=Replace_nulls(splitted_answers)
    
    return modified_values
    

##### ________________________________________________________________________________________________________________________

# Preprocessing

# Read data
raw_data=pd.read_csv(DATA_PATH)
data=raw_data.copy()

## Replace values

for col in REPLACE_COLS:
    data[col]=data[col].replace(REPLACE_DICT).astype(np.float32)
    
# we applied parsing by changing type to float32

## Split multiple values

object_cols = data.select_dtypes(include='object').columns.tolist()
for col in object_cols:
    
    data[col]=split_multianswers(data[col])

##### ________________________________________________________________________________________________________________________

# Visually verify results

idx=data.sample(1).index[0]

print(raw_data['LanguageWantToWorkWith'].iloc[idx])
print(data['LanguageWantToWorkWith'].iloc[idx])

idx=data.sample(1).index[0]

print(raw_data['DevType'].iloc[idx])
print(data['DevType'].iloc[idx])

idx=data.sample(1).index[0]

print(raw_data['YearsCode'].iloc[idx])
print(data['YearsCode'].iloc[idx])

print(raw_data['YearsCode'].loc[raw_data['YearsCode']=='Less than 1 year'])
print('_'*100) 
print(data['YearsCode'].loc[raw_data['YearsCode']=='Less than 1 year'])


##### ________________________________________________________________________________________________________________________

# Export data

data.to_pickle(EXPORT_PATH)

pk=pd.read_pickle(EXPORT_PATH)
pk.head()
