==> notebooks/05a_Prediction_notebook.ipynb <==
# constants 
MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_RUN_ID = "336de09f80cb4d16935f30f82e1e6ed8"

LOG_DATA_PKL = "data.pkl"
LOG_MODEL_PKL = "model.pkl"
LOG_METRICS_PKL = "metrics.pkl"

CLUSTERS_YAML_PATH = "../data/processed/features_skills_clusters_description.yaml"

# load packages
import os 
import sklearn
import pickle
import yaml

import pandas as pd

import mlflow
from mlflow.tracking import MlflowClient

## _Initialize_

### _1. Mlflow_

# convert path in the file to the format 
#-> 'C:\\Users\\DELL\\End2End_DS_Projects\\JobTechGuide\\notebooks\\..\\models\\mlruns\\830408098758583715\\336de09f80cb4d16935f30f82e1e6ed8\\artifacts'
# for windows

# Initialize client and experiment
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

run = mlflow.get_run(MLFLOW_RUN_ID)
artifacts_path = run.info.artifact_uri

### _Load model_

model_path = os.path.join(artifacts_path, LOG_MODEL_PKL)

# Load model
with open(model_path, "rb") as f:
    model = pickle.load(f)

model

artifacts_path

# Load data pkl
data_path  = os.path.join(artifacts_path, LOG_DATA_PKL)
with open(data_path, 'rb') as handle:
    data = pickle.load(handle)

data.keys()

# Unpack vars
features_names = pd.Series(data["features_names"])
targets_names  = pd.Series(data['targets_names'])
classifier = model['model_object']

classifier

### _Load skills Clusters_


# Load skills clusters
with open(CLUSTERS_YAML_PATH, "r") as stream:
    clusters_config = yaml.safe_load(stream)


# Reformat into data frame
molten_clusters = [(cluster_name, cluster_skill)
                   for cluster_name, cluster_skills in clusters_config.items()
                   for cluster_skill in cluster_skills]

clusters_df = pd.DataFrame(molten_clusters, columns=["cluster_name", "skill"])
clusters_df

## _Predict sample entry_

sample_skills = ['Scala', 'Hadoop', 'Python']

# Verify
pd.Series(sample_skills).isin(features_names)

### _1. Recreate cluster features_

sample_clusters = clusters_df.copy()
sample_clusters["sample_skills"] = sample_clusters["skill"].isin(sample_skills)
sample_clusters

cluster_features = sample_clusters.groupby("cluster_name")["sample_skills"].sum()
cluster_features

### _2. Create OneHotEncoded skills_

skills_names = features_names[~features_names.isin(cluster_features.index)]
skills_names

ohe_skills = pd.Series(skills_names.isin(sample_skills).astype(int).tolist(), 
                       index=skills_names)
ohe_skills

### _3. Combine features_

# Concat
features = pd.concat([ohe_skills,
                      cluster_features])


# Sort columns
features = features.loc[features_names]
features

### _4. Predict_

features.values

predictions = classifier.predict_proba([features.values])
predictions

positive_probs = [prob[0][1] for prob in predictions]
pd.Series(positive_probs, 
          index=targets_names).sort_values(ascending=False)
