==> notebooks/05b_Simulation_notebook.ipynb <==
MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_RUN_ID = "336de09f80cb4d16935f30f82e1e6ed8"

CLUSTERS_YAML_PATH = "../data/processed/features_skills_clusters_description.yaml"

import pandas as pd

import sys
sys.path.append('../')
from scripts.JobPrediction import *


## _Simulate additional skills_

### _1. Calculate base probabilities_

#current_skills = ['C', 'C++']
#current_skills = ['R', 'SQL', 'RStudio']
current_skills = ['Python', 'Pandas']

# Initialize model
model = JobPrediction(MLFLOW_TRACKING_URI, MLFLOW_RUN_ID, CLUSTERS_YAML_PATH)

# Calculate base probability
base_predictions = model.predict_jobs_probabilities(current_skills)
base_predictions.sort_values(ascending=False)

### _2. Calculate probabilities for new skills_

# Get possible new skills
all_skills = pd.Series(model.get_all_skills())
new_skills = all_skills[~all_skills.isin(current_skills)].copy()

current_skills in new_skills.values

simulated_results = []
for skill in new_skills:
    additional_skill_prob = model.predict_jobs_probabilities([skill] + current_skills)
    additional_skill_uplift = (additional_skill_prob - base_predictions) / base_predictions
    additional_skill_uplift.name = skill
    simulated_results.append(additional_skill_uplift)

simulated_results = pd.DataFrame(simulated_results)

simulated_results

### _Recommend_

target_job = 'Data scientist or machine learning specialist'

target_results = simulated_results[target_job].sort_values(ascending=False)
target_results

threshold = 1
recommendations = target_results[target_results > threshold].index.tolist()
recommendations

print("Your current skills: " + str(current_skills))
print("Your target job: " + str(target_job))
print("You might also consider learning: " + str(recommendations))
