==> notebooks/06_API_test.ipynb <==
# constants
API_URL = 'http://127.0.0.1:5000/'
PREDICTION_ENDPOINT = 'predict_jobs_probs'
SIMULATION_ENDPOINT = 'recommend_new_skills'

import pandas as pd
import requests
import json

### _Predict_

available_skills = ['R', 'SQL', 'RStudio']

predictions = requests.post(API_URL + PREDICTION_ENDPOINT,
                            data=json.dumps(available_skills),
                            headers={'content-type':'application/json'})

pd.Series(predictions.json()).sort_values(ascending=False)

### _Simulate_

target_job = 'Data scientist or machine learning specialist'

simulations = requests.post(API_URL + SIMULATION_ENDPOINT,
                            data=json.dumps({
                                'available_skills': available_skills,
                                'target_job': target_job
                            }),
                            headers={'content-type':'application/json'})

pd.Series(simulations.json()).sort_values(ascending=False)
